backend/pyproject.toml
[project]
name = "daily-tasks-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.16.5",
    "bcrypt>=4.0.0",
    "fastapi[standard]>=0.128.8",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
